flask_api/controllers/utils.py
# app/utils.py

import jwt
from functools import wraps
from flask import request, jsonify, current_app
from models import User
from passlib.hash import pbkdf2_sha256 as sha256
from datetime import datetime, timedelta

def token_required(func):
    @wraps(func)
    def decorated(*args, **kwargs):
        token = request.headers.get('Authorization')

        if not token:
            return jsonify({'message': 'Token is missing'}), 401

        try:
            data = jwt.decode(token, current_app.config['SECRET_KEY'], algorithms=['HS256'])
            current_user = User.query.filter_by(username=data['username']).first()
        except:
            return jsonify({'message': 'Token is invalid'}), 401

        return func(current_user, *args, **kwargs)

    return decorated
