flask_api/requirements.txt
Flask==3.0.3                  # Web frameworkFlask-Migrate==4.0.7         # DB migrationsFlask-SQLAlchemy==3.1.1      # ORM integrationSQLAlchemy==2.0.31           # Core database layerpsycopg2-binary==2.9.9       # PostgreSQL driveralembic==1.13.2              # Required by Flask-MigrateJinja2==3.1.4                # Templating engineMarkupSafe==2.1.5            # Jinja2 dependencyWerkzeug==3.0.3              # Flask server/utilsitsdangerous==2.2.0          # Flask session/key handlinggreenlet==3.0.3              # SQLAlchemy dependencyblinker==1.8.2               # Flask signalsclick==8.1.7                 # Flask CLIopenai==1.13.3               # OpenAI SDKpyjwt==2.8.0                 # JWT handlingpasslib==1.7.4               # Password hashingrequests==2.31.0             # HTTP requestsflask-cors==4.0.0            # CORS middleware# Unused/Dev-only — keep commented until confirmed neededpython-dotenv==1.0.1gunicorn# colorama==0.4.6# Mako==1.3.5# asgiref==3.8.1# sqlparse==0.5.0# typing_extensions==4.12.2# tzdata==2024.1